Revert using protobuf types with kolide api

+++ pkg/kolide-event-handler/types.go
@@ -14,7 +14,7 @@ type KolideEventHandler struct {
 }
 
 type KolideEventFailureData struct {
-	CheckId   uint64 `json:"check_id"`
+	CheckId   int    `json:"check_id"`
 	FailureId int    `json:"failure_id"`
 	Title     string `json:"title"`
 	DeviceId  int    `json:"device_id"`
